json parse for firefox

+++ BookmarkHighlighter/BookmarkHighlighter/Program.cs
@@ -1,30 +1,25 @@
 ﻿using BookmarkHighlighter.BookmarkParsers;
-using BookmarkHighlighter.Constants;
 using BookmarkHighlighter.JsonParser;
 using BookmarkHighlighter.JsWriters.GamesWriter;
 using BookmarkHighlighter.JsWriters.NexusModsWriter;
-using BookmarkHighlighter.Parser;
 
 // todo: make it so js file is only written once, first find all the links then write
 
-var bookmarksFilePath = @"X:\Misc Programs\Browser\Vivaldi\User Data\Default\Bookmarks";
-IJsonParser parser = new JsonParser(bookmarksFilePath);
+var bookmarksFilePath = @"/run/media/alex/storage/Browser Downloads/bookmarks-2024-07-28.json";
+IBookmarkParser parser = new FirefoxBackupParser(bookmarksFilePath);
 var bookmarkStructure = parser.GetBookmarkStructure();
 
-var outputPath = @"Highlighter Extension\GameHighlighter.js";
+var outputPath = @"Highlighter Extension/GameHighlighter.js";
 
-if (File.Exists(outputPath))
-{
-    File.Delete(outputPath);
-}
+if (File.Exists(outputPath)) File.Delete(outputPath);
 
 BookmarkParserBase gameParser = new GamesParser();
 BookmarkParserBase nexusModsParser = new NexusModsParser();
 
 var games = gameParser.Parse(bookmarkStructure);
-GamesJsWriter gamesJsWriter = new GamesJsWriter(outputPath);
+var gamesJsWriter = new GamesJsWriter(outputPath);
 gamesJsWriter.Write(games);
 
 var nexusMods = nexusModsParser.Parse(bookmarkStructure);
-NexusModsJsWriter nexusModsJsWriter = new NexusModsJsWriter(outputPath);
+var nexusModsJsWriter = new NexusModsJsWriter(outputPath);
 nexusModsJsWriter.Write(nexusMods);
